change: splitting CPU and GPU into different scripts to allow CPU docker to have an entrypoint

+++ machine_learning/docker/build_docker.sh
@@ -10,13 +10,15 @@ echo "Downloading model..."
 aws s3 cp --recursive s3://alchemab-ml/autoantibody_model temporary/autoantibody_model
 
 
+
+
 # Download predict script
 wget -O temporary/predict.py https://raw.githubusercontent.com/JessAlchemab/FinalYearProject/refs/heads/main/machine_learning/predict/predict.py
 wget -O temporary/analyse_metrics.py https://raw.githubusercontent.com/JessAlchemab/FinalYearProject/refs/heads/main/machine_learning/predict/analyse_metrics.py
 wget -O temporary/utils.py https://raw.githubusercontent.com/JessAlchemab/FinalYearProject/refs/heads/main/machine_learning/predict/utils.py
 wget -O temporary/aws_handler.py https://raw.githubusercontent.com/JessAlchemab/FinalYearProject/refs/heads/main/machine_learning/predict/aws_handler.py
 wget -O temporary/secrets_manager.py https://raw.githubusercontent.com/JessAlchemab/FinalYearProject/refs/heads/main/machine_learning/predict/secrets_manager.py
-wget -O temporary/handler.py 
+wget -O temporary/handler.py https://raw.githubusercontent.com/JessAlchemab/FinalYearProject/refs/heads/main/autoantibody_app/api/functions/classify-small/handler.py
 # Verify downloads
 if [ ! -d "temporary/autoantibody_model" ]; then
     echo "Error: model not downloaded"
@@ -38,16 +40,20 @@ if [ ! -f "temporary/secrets_manager.py" ]; then
     echo "Error: aws secrets manager not downloaded"
     exit 1
 fi
+if [ ! -f "temporary/handler.py" ]; then
+    echo "Error: classify small handler not downloaded"
+    exit 1
+fi
 
 echo "All assets downloaded successfully"
 
 
-docker build --no-cache . -t autoantibody_classifier
+docker build . -t autoantibody_classifier
 
 docker tag autoantibody_classifier 189545766043.dkr.ecr.eu-west-2.amazonaws.com/alchemab/autoantibody_classifier:latest
 docker push 189545766043.dkr.ecr.eu-west-2.amazonaws.com/alchemab/autoantibody_classifier:latest
 
-docker tag autoantibody_classifier 189545766043.dkr.ecr.eu-west-2.amazonaws.com/alchemab/autoantibody_classifier_gpu:latest
-docker push 189545766043.dkr.ecr.eu-west-2.amazonaws.com/alchemab/autoantibody_classifier_gpu:latest
+# docker tag autoantibody_classifier 189545766043.dkr.ecr.eu-west-2.amazonaws.com/alchemab/autoantibody_classifier_gpu:latest
+# docker push 189545766043.dkr.ecr.eu-west-2.amazonaws.com/alchemab/autoantibody_classifier_gpu:latest
 
 rm -rf temporary
